backend/requirements.txt
django==5.1.4
psycopg2-binary==2.9.10
django-bootstrap5==24.3
python-dotenv==1.0.1
pillow==11.1.0
django-cleanup==9.0.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djoser==2.3.1